+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_cifs_acl.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2018-2025, NetApp, Inc
+# (c) 2018-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
@@ -8,6 +8,7 @@ __metaclass__ = type
 
 DOCUMENTATION = '''
 author: NetApp Ansible Team (@carchi8py) <ng-ansible-team@netapp.com>
+short_description: NetApp ONTAP manage cifs-share-access-control
 description:
   - "Create or destroy or modify cifs-share-access-controls on ONTAP"
 extends_documentation_fragment:
@@ -48,8 +49,31 @@ options:
     type: str
     choices: [windows, unix_user, unix_group]
     version_added: 21.17.0
-short_description: NetApp ONTAP manage cifs-share-access-control
 
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.4.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 '''
 
 EXAMPLES = """
@@ -116,10 +140,12 @@ class NetAppONTAPCifsAcl:
             permission=dict(required=False, type='str', choices=['no_access', 'read', 'change', 'full_control']),
             type=dict(required=False, type='str', choices=['windows', 'unix_user', 'unix_group']),
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
             required_if=[
-                ('state', 'present', ['permission'])
+                ('state', 'present', ['permission']),
+                ('use_lambda', True, ['lambda_config'])
             ],
             supports_check_mode=True
         )
@@ -130,6 +156,8 @@ class NetAppONTAPCifsAcl:
         self.use_rest = self.rest_api.is_rest()
 
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if netapp_utils.has_netapp_lib() is False:
                 self.module.fail_json(msg=netapp_utils.netapp_lib_is_required())
             self.server = netapp_utils.setup_na_ontap_zapi(module=self.module, vserver=self.parameters['vserver'])